Bugfix: The message order is broken before sending to Inbound channel
The original code uses 25 goroutines to parse messages in parallel after
receiving them from the connection, and then send the result to Inbound
channel. It is possible that the Inbound channel consumer receives a
message earlier than another one which is sent later from OVS side,
because it is parsed more quickly. This implementation introduces an
issue that the order of the Multipart reply messages in one transaction
is broken.

The solution is to ensure the messages using the same transaction ID are
parsed by the same goroutine in the order they are received.

Signed-off-by: wenying <wenyingd@vmware.com>