Support continuous register mask for MatchField
For NXReg, OvS will accept value/mask on arbitrary bitwise. If only using
range[start:end], it is hard to combine discontinuous value/mask,
such as, to merge reg0[0:0] with reg0[2:2], with range method,
we will pass NXRange[0:2] and it will generate mask 0x7, but actually,
the mask is 0x5.

In this patch, we add a new API (NewRegMatchFieldWithMask) to support
arbitrary bitwise mask for register value/mask.

Signed-off-by: Jinjun Gao <gjinjun@gmail.com>