Bugfix: incorrect calculation on IP packet header length
This change is to resolve this issue: the header option length in an IP packet
header is not counted.

There are at least 20 bytes in the IP packet header. Hence, if the IHL value is
less than 5, 5 is used to count IP header length, otherwise it should use the IHL
value directly.

Signed-off-by: wenyingd <wenyingd@vmware.com>