Improve log level of verbose loggings
The OpenFlow messages are rather long and flooding the log file. In most
cases developers are not interested in the data of the messages while
V(4) is a widely used level for application logs.

The patch improves their log levels to 7, removes some redundant logs,
and fixes a log format issue in stream.go, which messed up log files.

Signed-off-by: Quan Tian <qtian@vmware.com>