Add support to set selection_method in GroupMod
Add support for setting selection_method, selection_method_param and fields in a
GroupMod message, the supported selection_method including "hash" and "dp_hash".
When using "dp_hash", fields is supposed to be empty. "fields" could use a mask
value or not, all bits in the value should be 1 if no mask is specified.

Signed-off-by: wenyingd <wenyingd@vmware.com>